The Pictures show CC instead of C
Hello,
I've followed the description and run into the issue that the C-Cross-Compiler is not detected, because in the pictures shown, at least for cross compilers, it's not "CC" but "C". It would be good if you could update the pictures to use the "C"-Definitions instead of "CC". Not sure if this applies to all compiler-definitions or just newer Cmake-Extension packages? Or is my CMake-Extension too old?